Refactor generic geo and add not-implemented model test
