+++ test/geocel/GenericGeoTestBase.t.hh
@@ -17,6 +17,7 @@
 #include "geocel/GeantGeoParams.hh"
 
 #include "CheckedGeoTrackView.hh"
+#include "GenericGeoResults.hh"
 #include "TestMacros.hh"
 #include "UnitUtils.hh"
 
@@ -324,8 +325,6 @@ auto GenericGeoTestBase<HP>::volume_stack(Real3 const& pos)
     -> VolumeStackResult
 {
     auto geo = this->make_geo_track_view(pos, Real3{0, 0, 1});
-    auto const& geo_params = *this->geometry();
-    auto const& vol_inst = geo_params.volume_instances();
 
     auto level = geo.level();
     if (!level)
@@ -335,31 +334,17 @@ auto GenericGeoTestBase<HP>::volume_stack(Real3 const& pos)
     std::vector<VolumeInstanceId> inst_ids(level.get() + 1);
     geo.volume_instance_id(make_span(inst_ids));
 
-    VolumeStackResult result;
-    result.volume_instances.resize(inst_ids.size());
-    result.replicas.assign(inst_ids.size(), -1);
-    for (auto i : range(inst_ids.size()))
-    {
-        auto vi_id = inst_ids[i];
-        if (!vi_id)
-        {
-            result.volume_instances[i] = "<null>";
-            continue;
-        }
-        auto const& label = vol_inst.at(vi_id);
-        if (auto phys_inst = geo_params.id_to_geant(vi_id))
-        {
-            if (phys_inst.replica)
-            {
-                result.replicas[i] = phys_inst.replica.get();
-            }
-        }
-        // Only write extension if not a replica, because Geant4
-        // effectively gives multiple volume instances the same name+ext
-        result.volume_instances[i] = !result.replicas[i] ? to_string(label)
-                                                         : label.name;
-    }
-    return result;
+    return VolumeStackResult::from_span(*this->geometry(), make_span(inst_ids));
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Get the model input from the geometry.
+ */
+template<class HP>
+auto GenericGeoTestBase<HP>::model_inp() const -> ModelInpResult
+{
+    CELER_NOT_IMPLEMENTED("model_inp");
 }
 
 //---------------------------------------------------------------------------//
