+++ test/geocel/GenericGeoTestInterface.hh
@@ -17,72 +17,14 @@
 
 class G4VPhysicalVolume;
 
-// DEPRECATED: remove in v0.7
-#define EXPECT_RESULT_EQ(EXPECTED, ACTUAL) EXPECT_REF_EQ(EXPECTED, ACTUAL)
-#define EXPECT_RESULT_NEAR(EXPECTED, ACTUAL, TOL) \
-    EXPECT_REF_NEAR(EXPECTED, ACTUAL, TOL)
-
 namespace celeritas
 {
 namespace test
 {
-
-class GenericGeoTestInterface;
-
 //---------------------------------------------------------------------------//
-struct GenericGeoTrackingResult
-{
-    std::vector<std::string> volumes;
-    std::vector<std::string> volume_instances;
-    std::vector<real_type> distances;  //!< [cm]
-    std::vector<real_type> halfway_safeties;  //!< [cm]
-    // Locations the particle had a very tiny distance in a volume
-    std::vector<real_type> bumps;  //!< [cm * 3]
-
-    void print_expected();
-};
-
-struct GenericGeoTrackingTolerance
-{
-    real_type distance{0};
-    real_type safety{0};
-
-    static GenericGeoTrackingTolerance
-    from_test(GenericGeoTestInterface const&);
-};
-
-//---------------------------------------------------------------------------//
-struct GenericGeoVolumeStackResult
-{
-    std::vector<std::string> volume_instances;
-    std::vector<int> replicas;
-
-    void print_expected();
-};
-
-//---------------------------------------------------------------------------//
-::testing::AssertionResult IsRefEq(char const* expected_expr,
-                                   char const* actual_expr,
-                                   char const* tol_expr,
-                                   GenericGeoTrackingResult const& expected,
-                                   GenericGeoTrackingResult const& actual,
-                                   GenericGeoTrackingTolerance const& tol);
-
-//---------------------------------------------------------------------------//
-inline ::testing::AssertionResult
-IsRefEq(char const* expected_expr,
-        char const* actual_expr,
-        GenericGeoTrackingResult const& expected,
-        GenericGeoTrackingResult const& actual)
-{
-    return IsRefEq(expected_expr, actual_expr, "default", expected, actual, {});
-}
-
-//---------------------------------------------------------------------------//
-::testing::AssertionResult IsRefEq(char const* expected_expr,
-                                   char const* actual_expr,
-                                   GenericGeoVolumeStackResult const& expected,
-                                   GenericGeoVolumeStackResult const& actual);
+struct GenericGeoTrackingResult;
+struct GenericGeoVolumeStackResult;
+struct GenericGeoModelInp;
 
 //---------------------------------------------------------------------------//
 /*!
@@ -101,6 +43,7 @@ class GenericGeoTestInterface
     //! \name Type aliases
     using TrackingResult = GenericGeoTrackingResult;
     using VolumeStackResult = GenericGeoVolumeStackResult;
+    using ModelInpResult = GenericGeoModelInp;
     using SPConstGeoInterface = std::shared_ptr<GeoParamsInterface const>;
     //!@}
 
@@ -114,9 +57,12 @@ class GenericGeoTestInterface
     //!@}
 
     //!@{
-    // Obtain the "touchable history" at a point
+    //! Obtain the "touchable history" at a point
     virtual VolumeStackResult volume_stack(Real3 const& pos_cm) = 0;
     //!@}
+
+    //! Get the model input from the geometry
+    virtual ModelInpResult model_inp() const = 0;
 
     //! Get the label for this geometry: Geant4, VecGeom, ORANGE
     virtual std::string_view geometry_type() const = 0;
